requirements.txt
absl-py==0.3.0
altgraph==0.16.1
appdirs==1.4.3
astor==0.7.1
bleach==1.5.0
cefpython3==66.0
certifi==2018.4.16
cffi==1.11.5
cgroupspy==0.1.6
chardet==3.0.4
contextlib2==0.5.5
coverage==4.5.1
Django==2.1.1
fs==2.0.21
future==0.17.1
gast==0.2.0
h5py==2.8.0
html5lib==0.9999999
idna==2.6
Jinja2==2.10
Keras==2.1.6
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
lxml==4.2.1
macholib==1.11
Markdown==2.6.11
MarkupSafe==1.0
msgpack==0.5.6
nsenter==0.2
numpy==1.14.3
pandas==0.23.0
pathlib==1.0.1
pefile==2018.8.8
Pillow==5.1.0
protobuf==3.6.1
pyasn1==0.4.4
pycparser==2.18
PyInstaller==3.4
PyNaCl==1.2.1
pyotp==2.2.6
PyQRCode==1.2.1
python-bitcoinrpc==1.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.18.4
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.1.0
shortuuid==0.5.0
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
tqdm==4.23.3
urllib3==1.22
zstandard==0.9.1
python-prctl==1.7
